refactor: split `extraReducer`s  from slice files
